I had an Angular itch to scratch.
Angular isn't something I've really gotten to use before and I wanted to
try it out, so I decided to rework this IGN example using Angular.

I am so impressed how easy this was an how little code it took compared
to jQuery. Angular is DEFINITELY the way to go for working with API's
and client-side rendering.

The trickiest part was the cross-origin API call. Unlike jQuery, you
have to include the callback_function query string because Angular
expects a callback function to be returned and will throw an error if it
gets plain JSON.

(It's not complete, but you'll get the idea.)